+++ frontend_react/package.json
@@ -25,6 +25,7 @@
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
+    "react-material-ui-carousel": "^2.3.11",
     "react-redux": "^7.2.5",
     "react-scripts": "^3.4.1",
     "react-twitter-embed": "^3.0.3",
